+++ redis_metrics/tests/__init__.py
@@ -1,4 +1,5 @@
 from .forms import TestAggregateMetricForm, TestMetricCategoryForm
 from .models import TestR
+from .templatetags import TestTemplateTags
 from .views import TestViews
 from .utils import TestUtils