+++ HISTORY.rst
@@ -3,7 +3,7 @@ History
 
 0.6.0 (in progess)
 ++++++++++++++++++
-
+- Added a ``gauge`` template tag.
 - Added methods & management commands to delete metrics & gauges
 
 0.5.1 (2013-05-18)